requirements.txt
pywinauto==0.6.9
pyautogui==0.9.54
pynput==1.8.1
Pillow==11.3.0
python-dotenv==1.2.1
